refactor: move util inside packages
