+++ packages/kit/src/imports.ts
@@ -2,7 +2,7 @@ import type { Import } from 'unimport'
 import type { ImportPresetWithDeprecation } from '@nuxt/schema'
 import { useNuxt } from './context'
 import { assertNuxtCompatibility } from './compatibility'
-import { toArray } from '../../../utils'
+import { toArray } from './utils'
 
 export function addImports (imports: Import | Import[]) {
   assertNuxtCompatibility({ bridge: true })
