+++ packages/nuxt/src/pages/runtime/plugins/router.ts
@@ -24,7 +24,7 @@ import _routes from '#build/routes'
 import routerOptions from '#build/router.options'
 // @ts-expect-error virtual file
 import { globalMiddleware, namedMiddleware } from '#build/middleware'
-import { toArray } from '../../../../../../utils'
+import { toArray } from '../../../utils'
 
 // https://github.com/vuejs/router/blob/4a0cc8b9c1e642cdf47cc007fa5bbebde70afc66/packages/router/src/history/html5.ts#L37
 function createCurrentLocation (
