+++ packages/webpack/src/utils/index.ts
@@ -1,0 +1,3 @@
+export function toArray<T> (value: T | T[]): T[] {
+  return Array.isArray(value) ? value : [value]
+}
